refactor: review comments, rename stuff, move tests to blob package
